Removed unneeded casts

+++ Server/Items/VirtualHair.cs
@@ -171,11 +171,11 @@ namespace Server
 
             int faceSerial = FaceInfo.FakeSerial(parent);
 
-            m_Stream.Write((int)faceSerial);
+            m_Stream.Write(faceSerial);
             m_Stream.Write((short)parent.FaceItemID);
             m_Stream.Write((byte)0);
             m_Stream.Write((byte)Layer.Face);
-            m_Stream.Write((int)parent.Serial);
+            m_Stream.Write(parent.Serial);
             m_Stream.Write((short)hue);
         }
     }
@@ -203,7 +203,7 @@ namespace Server
         public RemoveFace(Mobile parent)
             : base(0x1D, 5)
         {
-            m_Stream.Write((int)FaceInfo.FakeSerial(parent));
+            m_Stream.Write(FaceInfo.FakeSerial(parent));
         }
     }
 }
